Getting first stages of logging / error handling
Started the first stages of logging and errorhandling to help work on developement. Currently the log files are just saved in /logs and are named based off of the system time. The log appending functionality for now simply takes the most recent file by file name and appends (or will append when the work is done) the appropriate information to the file.